+++ src/hb_grid/hb_grid_request_loginuser.c
@@ -7,6 +7,8 @@
 #include "hb_json/hb_json.h"
 #include "hb_utils/hb_base16.h"
 
+#include <string.h>
+
 int hb_grid_request_loginuser( struct evhttp_request * _request, struct hb_grid_process_handle_t * _handle, char * _response, size_t * _size, const char * _pid )
 {
     hb_node_loginuser_in_t in_data;
@@ -87,7 +89,7 @@ int hb_grid_request_loginuser( struct evhttp_request * _request, struct hb_grid_
         hb_token_base16_encode( out_data.token, token16 );
 
         size_t response_data_size = sprintf( _response, "{\"code\": 0, \"token\": \"%.*s\"}"
-            , sizeof( token16 )
+            , (int)sizeof( token16 )
             , token16
         );
